subspace_archiving

Module archiver

Structs

ArchiveBlockOutcome

The outcome of adding a block to the archiver.

Archiver

Block archiver for Subspace blockchain.

NewArchivedSegment

Newly archived segment as a combination of segment header and corresponding archived history segment containing pieces

Enums

ArchiverInstantiationError

Archiver instantiation error

Segment

Segment represents a collection of items stored in archival history of the Subspace blockchain

SegmentItem

Kinds of items that are contained within a segment