Type Definition subspace_archiving::archiver::BlockArchiver

pub type BlockArchiver = Archiver<BlockArchiverState>;

Block archiver for Subspace blockchain.

It takes new confirmed (at K depth) blocks and concatenates them into a buffer, buffer is sliced into segments of RECORDED_HISTORY_SEGMENT_SIZE size, segments are sliced into source records of RECORD_SIZE, records are erasure coded, Merkle Tree is built over them, and with Merkle Proofs appended records become pieces that are returned alongside corresponding root block header.

Implementations

impl BlockArchiver

pub fn with_initial_state<B: AsRef<[u8]>>(
record_size: usize,
segment_size: usize,
root_block: RootBlock,
encoded_block: B,
object_mapping: BlockObjectMapping
) -> Result<Self, ArchiverInstantiationError>

Create a new instance of block archiver with initial state in case of restart.

block corresponds to last_archived_block and will be processed accordingly to its state.

pub fn last_archived_block_number(&self) -> Option<u32>

Get last archived block if there was any

pub fn add_block(
&mut self,
bytes: Vec<u8>,
object_mapping: BlockObjectMapping
) -> Vec<ArchivedSegment>

Adds new block to internal buffer, potentially producing pieces and root block headers