Type Definition subspace_archiving::archiver::ObjectArchiver

pub type ObjectArchiver = Archiver<ObjectArchiverState>;

Object archiver for Subspace blockchain.

It takes pre-genesis objects and concatenates them into a buffer, buffer is sliced into segments of RECORDED_HISTORY_SEGMENT_SIZE size, segments are sliced into source records of RECORD_SIZE, records are erasure coded, Merkle Tree is built over them, and with Merkle Proofs appended records become pieces that are returned alongside corresponding root block header.

Object archiver is only used pre-genesis and should be turned into block archiver for later use.

Implementations

impl ObjectArchiver

pub fn add_object(&mut self, object: Vec<u8>) -> Vec<ArchivedSegment>

Adds new object to internal buffer, potentially producing pieces and root block headers

pub fn into_block_archiver(self) -> BlockArchiver