Module subspace_archiving::archiver

Structs

Archiver

Block archiver for Subspace blockchain.

NewArchivedSegment

Newly archived segment as a combination of segment header hash, segment index and corresponding archived history segment containing pieces

Enums

ArchiverInstantiationError

Archiver instantiation error

Segment

Segment represents a collection of items stored in archival history of the Subspace blockchain

SegmentItem

Kinds of items that are contained within a segment

Functions

is_piece_valid

Validate witness embedded within a piece produced by archiver

is_record_commitment_hash_valid

Validate witness for record commitment hash produced by archiver